Add interoperability with python code (tentative)
Utils function for visual assesment (e.g. sampling runs) were already implemented
(and debugged!) in python, so the appropriate code is imported as python module and
run as it is. Code can be found in `python_interop` directory.

TODO: the heavy lifting pipeline (load frames, multifoveal filtering, save video)
still has to be implemented. This must necessarily be implemented in Matlab.